feat: Implement contact editing functionality with modal and API integration
